Added Multiple Responses

+++ MyuTwitchBot/Program.cs
@@ -8,20 +8,10 @@ namespace MyuTwitchBot
     {
         static void Main(string[] args)
         {
-            Boolean exit = false;
-            if (!File.Exists("./config.json"))
-            {
-                File.WriteAllText("./config.json", JsonConvert.SerializeObject(new Configuration(), Formatting.Indented));
-                exit = true;
-            }
-            if (!File.Exists("./commands.json"))
-            {
-                File.WriteAllText("./commands.json", JsonConvert.SerializeObject(new Command[] { new Command() }, Formatting.Indented));
-                exit = true;
-            }
-            if (exit) Environment.Exit(1);
-
+            if (!File.Exists("./config.json")) File.WriteAllText("./config.json", JsonConvert.SerializeObject(new Configuration(), Formatting.Indented));
+            if (!File.Exists("./commands.json")) File.WriteAllText("./commands.json", JsonConvert.SerializeObject(new Command[] { new Command() }, Formatting.Indented));
             Bot twitchBot = new Bot(JsonConvert.DeserializeObject<Configuration>(File.ReadAllText("./config.json")));
+            Console.WriteLine("1");
             Console.ReadLine();
         }
     }
